+++ internal/blockchain/ethereum/besu/besuCliqueConfig/besu/config.toml
@@ -32,11 +32,10 @@ permissions-nodes-config-file-enabled=true
 permissions-nodes-config-file="/config/besu/permissions_config.toml"
 
 # bootnodes
-bootnodes=["enode://c1979a8a48693db804316b5acebe35e11731e1fb1c9c21ff7268ab25db6f6e03390a429b83cf0ec0865a7205f2669ec1ace652a3def11e2e01571c74939cbe22@172.16.239.11:30303"]
+bootnodes=["enode://c1979a8a48693db804316b5acebe35e11731e1fb1c9c21ff7268ab25db6f6e03390a429b83cf0ec0865a7205f2669ec1ace652a3def11e2e01571c74939cbe22@validator1:30303"]
 
 
 # Discovery at boot is set to a list of static files, but will also discover new nodes should they be added
 # static nodes
 static-nodes-file="/config/besu/static-nodes.json"
 discovery-enabled=true
-
