+++ internal/blockchain/ethereum/besu/besuCliqueConfig/besu_mem1_def.sh
@@ -1,11 +1,8 @@
 #!/bin/bash
-ip=$(hostname -i)
-
- 
 for i in $(seq 1 100) 
 do
 	set -e
-	response="$(curl -0 -m 10 172.16.239.26:9000/upcheck)"
+	response="$(curl -0 -m 10 member1tessera:9000/upcheck)"
 	if [[ "I'm up!" == $response ]];
 	then break
 	else
@@ -18,8 +15,9 @@ done
 
 while [ ! -f "/opt/besu/public-keys/bootnode_pubkey" ]; do sleep 5; done ;
 /opt/besu/bin/besu \
+--Xdns-enabled=true \
+--Xdns-update-enabled=true \
 --config-file=/config/besu/config.toml \
---p2p-host=$ip \
 --genesis-file=/config/besu/CliqueGenesis.json \
 --node-private-key-file=/opt/besu/keys/key \
 --min-gas-price=0 \
