+++ internal/blockchain/ethereum/besu/besuCliqueConfig/ethsigner/ethsigner.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
 apt-get update
-yes | apt-get install curl
+apt-get install -y curl
 
 bin/ethsigner --chain-id=1337 --http-listen-host=0.0.0.0 --downstream-http-port=8555 --downstream-http-host=rpcnode multikey-signer --directory=/SignerConfig
