+++ internal/blockchain/ethereum/besu/besuCliqueConfig/validator_node_def.sh
@@ -1,11 +1,9 @@
 #!/bin/bash
-ip=$(hostname -i)
-echo "Here now ${BESU_CONS_API:-IBFT}"
-echo "IP is $ip"
 while [ ! -f "/opt/besu/public-keys/bootnode_pubkey" ]; do sleep 5; done ; \
 /opt/besu/bin/besu \
+--Xdns-enabled=true \
+--Xdns-update-enabled=true \
 --config-file=/config/besu/config.toml \
---p2p-host=$ip \
 --genesis-file=/config/besu/CliqueGenesis.json \
 --node-private-key-file=/opt/besu/keys/key \
 --min-gas-price=0 \
