+++ internal/tokens/erc1155/erc1155_provider.go
@@ -20,7 +20,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/hyperledger/firefly-cli/internal/constants"
 	"github.com/hyperledger/firefly-cli/internal/core"
 	"github.com/hyperledger/firefly-cli/internal/docker"
 	"github.com/hyperledger/firefly-cli/internal/log"
@@ -50,7 +49,6 @@ func (p *ERC1155Provider) FirstTimeSetup() error {
 
 func (p *ERC1155Provider) GetDockerServiceDefinitions() []*docker.ServiceDefinition {
 	serviceDefinitions := make([]*docker.ServiceDefinition, 0, len(p.Stack.Members))
-	netId := constants.NetID * 5
 	for i, member := range p.Stack.Members {
 		serviceDefinitions = append(serviceDefinitions, &docker.ServiceDefinition{
 			ServiceName: "tokens_" + member.ID,
@@ -71,11 +69,6 @@ func (p *ERC1155Provider) GetDockerServiceDefinitions() []*docker.ServiceDefinit
 					Test: []string{"CMD", "curl", "http://localhost:3000/api"},
 				},
 				Logging: docker.StandardLogOptions,
-				Networks: &docker.Network{
-					fmt.Sprintf("%s_default", p.Stack.Name): &docker.IPMapping{
-						IPAddress: fmt.Sprintf("172.16.239.%v", netId+i),
-					},
-				},
 			},
 		})
 	}
